backend/requirements.txt
Django==5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
django-cors-headers==4.3.1

# Financial APIs
yfinance==0.2.36
alpha_vantage==2.3.1
finnhub-python==2.4.18
twelvedata==1.2.11
fmp-python==0.1.5

# AI/ML
google-generativeai==0.3.2

# Async support
celery==5.3.6
redis==5.0.1

# Development
django-debug-toolbar==4.3.0

# Swagger documentation
drf-yasg==1.20.0
finnhub-python
